Update deck name verify
Will parse the deck name format of the Weekly Tournament in iDuel `11111111+Yuzurisa.ydk`. The name matches only when the deck name is exactly the same as the player's name (`11111111+Yuzurisa` in this example), the first part of the name (`11111111` here), or the second part (`Yuzurisa` here) .
I made this PR because many organizers of Weekly Tournament reported that in some cases, the deck name would match wrongly during the tournament. Please carefully review this, thanks!